requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dateutil>=2.8.2
python-multipart>=0.0.6
pymysql>=1.1.0
cryptography>=41.0.0
